Properly separate out long tests into new profile
