+++ seqware-ext-testing/pom.xml
@@ -217,6 +217,8 @@
 
     <profiles>
         <profile>
+            <!-- run integration tests (IT) that require postgres and tomcat -->
+            <!-- run extended integration tests (ET) that require Oozie -->
             <id>extITs</id>
             <build>
                 <plugins>
@@ -231,6 +233,32 @@
                                 <include>**/*IT.java</include>
                                 <include>**/ET*.java</include>
                                 <include>**/*ET.java</include>
+                            </includes>
+                        </configuration>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+        <profile>
+            <!-- run integration tests (IT) that require postgres and tomcat -->
+            <!-- run extended integration tests (ET) that require Oozie -->
+            <!-- run long integration tests (LT) that take more time than can be provided by Travis-CI -->
+            <id>allITs</id>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>org.apache.maven.plugins</groupId>
+                        <artifactId>maven-failsafe-plugin</artifactId>
+                        <version>${maven-failsafe-plugin.version}</version>
+
+                        <configuration>
+                            <includes>
+                                <include>**/IT*.java</include>
+                                <include>**/*IT.java</include>
+                                <include>**/ET*.java</include>
+                                <include>**/*ET.java</include>
+                                <include>**/LT*.java</include>
+                                <include>**/*LT.java</include>
                             </includes>
                         </configuration>
                     </plugin>
