+++ seqware-ext-testing/src/test/java/net/sourceforge/seqware/pipeline/tutorial/OldAdminTutorialSuiteLT.java
@@ -18,20 +18,17 @@ package net.sourceforge.seqware.pipeline.tutorial;
 
 import net.sourceforge.seqware.pipeline.plugins.ExtendedTestDatabaseCreator;
 import org.junit.BeforeClass;
-import org.junit.Ignore;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
 /**
- * This runs all the tests needed to test the Developer Tutorials. The test suite actually enforces order for us. We actually started with
- * built and installed bundles from step 5 of the user tutorials.
+ * This runs all the tests needed to test the Admin Tutorials. The test suite actually enforces order for us.
  * 
  * @author dyuen
  */
-@Ignore("see https://github.com/SeqWare/seqware/issues/324")
 @RunWith(Suite.class)
-@Suite.SuiteClasses(value = { DeveloperPhase1.class, DeveloperPhase2.class })
-public class OldDeveloperTutorialSuiteET extends TutorialSuite {
+@Suite.SuiteClasses(value = { AdminPhase1.class })
+public class OldAdminTutorialSuiteLT extends TutorialSuite {
     @BeforeClass
     public static void resetDatabase() {
         ExtendedTestDatabaseCreator.resetDatabaseWithUsers();
